feat: AIコーチ機能をWhisper API対応に変更し、音声再生機能を改善
- Web Speech APIからWhisper APIへの移行
- MediaRecorder APIを使用した音声録音機能の実装
- OpenAI TTS APIを使用した高品質な音声再生機能
- 音声の重複再生を防止する機能
- 録音停止ボタンの動作改善
- duration_secondsをintからfloatに変更（バックエンドスキーマ）
- エラーハンドリングとデバッグログの追加

+++ backend/app/models/schemas.py
@@ -86,7 +86,7 @@ class WhisperTranscribeRequest(BaseModel):
     """Whisper文字起こしリクエスト"""
     audio_data: str  # base64エンコードされた音声データ
     session_id: str
-    duration_seconds: int
+    duration_seconds: float  # 録音時間（秒）
 
 
 class WhisperTranscribeResponse(BaseModel):
